Fix s2s probe to ejabberd(?) instances
The stream negotiation is stuck for some reason there, while it does
work with Prosody.

By attaching a simple SASL offered check (i.e. add the SASL stream
feature to the mellium negotiator), we can circumvent this
situation. This is because the CheckSASLOffered thing claims that
the stream is ready after doing nothing :).