Use encoding API for sending register IQs
While looking at the code mentioned in mellium/xmpp#33 I noticed that
this project was using the "Send" method to send Register IQs instead
of "SendIQ". Because Send doesn't block and wait for the IQ response,
what we really need to remove the need for a TokenReader is to use the
Encode method which already exists. Updating to use Encode makes it
easier to send stanzas without having to write a large TokenReader
implementation, at the cost of efficiency (see mellium/xmpp#38).