Reuse XMPP sessions for ping probes
Without session reusal, the ping probes can become quite expensive
both for the prober and the system providing the probing account.
In such a case, each ping requires a complete TLS and SASL
handshake, potentially with many iterations of PBKDF2 for SCRAM,
eating a considerable amount of resources.

By enabling session sharing by default (with a switch to turn it
off), ping checks become faster and cheaper. The downside is that
the existing connection may break in interesting ways.
Health-checking and dealing with broken connections (which aren’t
detected by the OS already) is to be implemented.

This change introduces a breaking change in the config format, by
requiring all ping checks to refer to an account defined on the top
level instead of defining the account details inline; this is to
have the session sharing relationships between ping checks explicit.