Implement client account healthchecks
Healthchecks are used to force-terminate an existing XMPP session
in case it goes bad. Healthchecks are simply XEP-0199 pings
against the account’s domain.

If the domain answers the ping (with an error or a success, it
doesn’t matter), the session is considered alive.

Healthchecks are triggered whenever a probe using the account
fails for a reason which could be caused by a bad connection, for
example a ping timeout.

As usual, if a client does not have a session, a new session
will be established for the next probe using the account. Thus,
when a session failed once, it will be automatically re-established
for the next probe.