client: correctly pass context (including timeout) to connection
Otherwise ... Well, let’s suppose you have a ping account where the
provider suddenly vanishes with a connection timeout. And it
doesn’t come back.

And then you have a Prometheus which is scraping that. Multiple
times per minute. And then it accumulates and gets worse until
whatever default connection timeout kicks in (hint: it’s not going
to be the 10s of prometheus, or the 60s of your scrape interval).

And the exporter will be collecting more and more open file
descriptors. And then boom.