client: use a try-able semaphore instead of a mutex for the session
The key advantage is that if the session creation gets stuck for
any reason, the other probes which attempt to use the session will
still eventually cancel and return a result to the caller.

This is of importance as a matter of resilience; if other probes
get stuck on a shared session (because they wait for the lock
forever), we end up consuming more and more filedescriptors (and
also spawning more and more goroutines) until eventually
everything breaks.