Change logic for detecting certificate expiry
The previous logic ignored to which chain a certificate belongs. That
was especially fun with the Let's Encrypt root certificate rollover [1].

The problem here was (or, still is) that for any Let's Encrypt cert
using the compatibility intermediate certificate, there are two valid
chains (until next week): One which uses

        CN=DST Root CA X3,O=Digital Signature Trust Co.

and one which stops at

        CN=ISRG Root X1,O=Internet Security Research Group,C=US

(which is, by now, in many root stores).

The old logic would return the expiry of the CN=DST Root CA X3
certificate only, which pointed at next week. That was completely
ignoring that the root store had another way to verify the certificate,
by stopping at CN=ISRG Root X1 which is also part of the root store.

That caused a lot of false alarms.

To avoid this situation in the future, we now take into account the
different possible chains: we return two timestamps now, one for the
*next* chain to expire (which will not be used for alerting in general)
and one for the *last* chain to expire (this will in almost all cases be
the time when the end entity certificate expires, unless something is
really f*ed up).

The metric name is now misleading, but I didn't want to break that for
now.